backend/requirements.txt
asgiref==3.5.0
brotlipy==0.7.0
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1631636293358/work
chardet @ file:///Users/runner/miniforge3/conda-bld/chardet_1610093454858/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1626371162869/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1602866480661/work
conda==4.10.3
conda-package-handling @ file:///Users/runner/miniforge3/conda-bld/conda-package-handling_1618231726924/work
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography_1633983374178/work
distlib==0.3.4
Django==4.0.2
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
djangorestframework==3.13.1
filelock==3.6.0
firebase-admin==5.2.0
gcloud==0.17.0
google-api-core==2.6.0
google-api-python-client==2.39.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.3.1
googleapis-common-protos==1.55.0
grpcio==1.44.0
grpcio-status==1.44.0
httplib2==0.20.4
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1609836280497/work
jws==0.1.3
mamba @ file:///Users/runner/miniforge3/conda-bld/mamba_1632770424175/work
msgpack==1.0.3
oauth2client==3.0.0
packaging==21.3
Pillow==9.0.1
pipenv==2022.1.8
platformdirs==2.5.1
proto-plus==1.20.3
protobuf==3.20.0rc1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat @ file:///Users/runner/miniforge3/conda-bld/pycosat_1624996792650/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
pycryptodome==3.4.3
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1633192417276/work
pyparsing==3.0.7
Pyrebase==3.0.27
PySocks @ file:///Users/runner/miniforge3/conda-bld/pysocks_1610291460269/work
python-jwt==2.0.1
pytz==2021.3
requests==2.27.1
requests-toolbelt==0.7.0
rsa==4.8
ruamel-yaml-conda @ file:///Users/runner/miniforge3/conda-bld/ruamel_yaml_1624999868738/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sqlparse==0.4.2
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1632160078689/work
uritemplate==4.1.1
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1632350318291/work
virtualenv==20.13.1
virtualenv-clone==0.5.7
